Added a few parameters to the DebPlugin:
  * libTargetDirectory - default usr/share/java/${project.group.id} 

From what I could see, most Debian packages install jars in usr/share/java - so libcommons-lang-java, puts a jar file in 
usr/share/java/commons-lang-2.1.jar.   While we could probably get away with 
putting JARs in this same directory, it didn't feel appropriate to just use the artifactId and the version number.  I've added another directory that captures the groupId.   This is a break from what 
the original DebPlugin did, the originial plugin put jars in usr/local/jars which isn't compatible with what the Debian project.  (And, I know, it isn't that we have to be Debian compliant, but why 
not start with compatibility and let people override if they need to.)

  * includeJavaDocs - Added a boolean flag (default true) 

allows you to specify whether or not the deb plugin should bundle Javadoc in the deb file.  Again, I added this after looking at the DEB 
packages distributed by both the Debian and Ubuntu projects.   All of them have Javadoc available in /usr/share/doc.

  * docSourceDirectory - default ${basedir}/site/apidoc

Allows you to specify where the deb plugin to copy Javadoc from.  This only works if you put associate with javadoc goal with the package stage.

  * docTargetDirectory - default usr/share/doc/${project.groupId}/${project.artifactId}/api

Again this directory is consistent with how debian already packages libraries like the commons libraries.