[MVERSIONS-89] Fixes the NPE...
o of course the report is not quite functional when dealing with a version range... 

o need to think a bit about how to fix the report for version ranges... it seems that dependency resolution is not happening for the report, so we do not get the resolved range. I don;t want to reproduce all the range resolution code, so I need to find a way to get the resolved range when generating the report.