MVERSIONS-249: implemented updatePropertyVersions
If the property is defined, the properties defining a  version matching the dependencies of the remote POM are updated. If a particular property defines the version of several artifacts, all artifacts should be defined by the remote POM and have one and only one version. In other cases, the property is not updated.